requirements.txt
Flask
gunicorn
google-generativeai
PyGithub